Fixed the backend-frontend communication. Changed backend services, controller, entity, dto to use List instead of Set. - probably not a proper way since it doesnt matter cos JSON gets serialized and deserialized regardless but my dumbass didnt figure out how to make the front end pick the actor and genres data as Array... All the service classes have methods to convert to dto and entity to sort out separation concerns? Set up the Movie entity to have JsonManagedReference and JsonBackReference for control between the entities as movie is the main and the actor and genre are secondary. Gonna have to examine it deeper how each part interacts with eachother. On the front end: Added some back buttons for the actor and genre pages Beautified everything to look alright.

+++ angular-frontend/src/app/actors/actors.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Location } from '@angular/common';
 import { ActorService } from '../services/actor.service';
 import { Actor } from '../models/actor.model';
 
@@ -11,30 +12,44 @@ export class ActorsComponent implements OnInit {
   actors: Actor[] = [];
   newActor: string = '';
 
-  constructor(private actorService: ActorService) {}
+  constructor(
+    private actorService: ActorService,
+    private location: Location,
+  ) { }
 
   ngOnInit() {
     this.getActors();
   }
-
+  // This retrieves the actors when the component initializes.
   getActors(): void {
+    console.log('Fetching actors...');
     this.actorService.getActors()
-      .subscribe(actors => this.actors = actors);
+      .subscribe(actors => {
+        console.log('Actors fetched: ', actors);
+        this.actors = actors;
+      });
   }
 
+  // Method to handle deletion of actor
   delete(actor: Actor): void {
-    this.actors = this.actors.filter(h => h !== actor);
-    this.actorService.deleteActor(actor.id).subscribe();
+    if (confirm(`Are you sure you want to delete ${actor.name}?`)) {
+      this.actors = this.actors.filter(h => h !== actor);
+      this.actorService.deleteActor(actor.id).subscribe();
+    }
   }
-
+  
+  // Method to handle adding a new actor
   addActor(): void {
     if (this.newActor) {
       this.actorService.addActor(this.newActor).subscribe(actor => {
         this.actors.push(actor);
-        this.newActor = ''; // Reset the input after adding
+        this.newActor = '';
       });
     }
   }
 
-  // Add more methods for editing, adding actors if needed
+  // Method for the button to go back
+  goBack(): void {
+    this.location.back();
+  }
 }
